ai-core/pyproject.toml
[project]
name = "ai-core"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "filterpy>=1.4.5",
    "ipykernel>=7.1.0",
    "lapx>=0.9.3",
    "numpy>=2.4.0",
    "opencv-python>=4.11.0.86",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "scipy>=1.16.3",
    "streamlit>=1.52.2",
    "ultralytics>=8.3.243",
]
